ServidorMapaVirtual/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.2
cycler==0.10.0
defusedxml==0.7.0rc1
Django==3.1
django-braces==1.14.0
django-cors-headers==3.4.0
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
numpy==1.19.1
oauthlib==3.1.0
Pillow==7.2.0
pycparser==2.20
pyfcm==1.4.7
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
